requirements.txt
fastapi
uvicorn
streamlit
requests
numpy
faiss-cpu
sentence-transformers
pymupdf
python-dotenv
matplotlib
pandas
groq
python-multipart